vantage6-server/requirements.txt
appdirs==1.4.3
flask==1.1.1
Flask-RESTful==0.3.8
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-SocketIO==4.2.1
SQLAlchemy==1.3.15
iknl-flasgger==0.9.2.post1
schema==0.7.1
bcrypt==3.1.7
questionary==1.5.2
marshmallow==2.16.3
marshmallow-sqlalchemy==0.15.0
ipython==7.13.0
requests==2.23.0
psutil==5.7.0
gevent==20.9.0
flask-mail==0.9.1
Flask-Principal==0.4.0
kombu==5.2.2